requirements.txt
Flask==3.0.2
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.9
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.6.0
faiss-cpu==1.8.0
Flask-Login==0.6.3
